Added support to provide server name as an argument.

+++ vnc_operations/start_vncservers.bash
@@ -1,22 +1,25 @@
 #!/bin/bash
 
-# We need to start 8 servers on different DISPLAYS on console.sv.vt.edu
+if [ $# -ne 1 ]; then
+    echo "Enter the server on which to start VNCSERVER."
+    exit 1
+fi
 
-# First, kill all existing vncservers on viz1
+SERVER=$1
+
+# We need to start 8 servers on different DISPLAYS on $SERVER
+
+# First, kill all existing vncservers on $SERVER
 KILL_VNCSERVER="pkill -f vnc"
-#echo $KILL_VNCSERVER
 
-ssh console.sv.vt.edu "$KILL_VNCSERVER"
+ssh $SERVER "$KILL_VNCSERVER"
 
-#: <<'END'
-# Start 8 vncservers on console
 START_VNCSERVER="vncserver -geometry 12560x1600"
 set -ex
 for i in {1..8}; do
-    if echo $START_VNCSERVER :$i | ssh console.sv.vt.edu bash; then
-        echo "VNC server successfully started on console.sv.vt.edu:$i"
+    if echo $START_VNCSERVER :$i | ssh $SERVER bash; then
+        echo "VNC server successfully started on $SERVER:$i"
     else
-        echo "Failed to start VNC server on console.sv.vt.edu:$i"
+        echo "Failed to start VNC server on $SERVER:$i"
     fi
 done
-#END
